requirements.txt
discord.py[voice]==1.5.1
PyAudio==0.2.11